feat: `fmt::Display` impl for `Pattern`. (#301)
This way the original pattern can be reproduced on the fly without
actually storing it, saving one allocation.
