+++ git-glob/src/pattern.rs
@@ -1,5 +1,6 @@
 use bitflags::bitflags;
 use bstr::{BStr, ByteSlice};
+use std::fmt;
 
 use crate::{pattern, wildmatch, Pattern};
 
@@ -140,5 +141,21 @@ impl Pattern {
                 }
             }
         }
+    }
+}
+
+impl fmt::Display for Pattern {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        if self.mode.contains(Mode::NEGATIVE) {
+            "!".fmt(f)?;
+        }
+        if self.mode.contains(Mode::ABSOLUTE) {
+            "/".fmt(f)?;
+        }
+        self.text.fmt(f)?;
+        if self.mode.contains(Mode::MUST_BE_DIR) {
+            "/".fmt(f)?;
+        }
+        Ok(())
     }
 }
